# api/docs/swagger.yaml
components:
  schemas:
...
    github_com_replicatedhq_kotskinds_multitype.BoolOrString:
      type: object
    types.APIError:
      properties:
        errors:
          items:
            $ref: '#/components/schemas/types.APIError'
          type: array
          uniqueItems: false
        field:
          type: string
        message:
          type: string
        status_code:
          type: integer
      type: object
    types.AppConfig:
      properties:
        groups:
          items:
            $ref: '#/components/schemas/v1beta1.ConfigGroup'
          type: array
          uniqueItems: false
      type: object
    types.AppConfigValue:
      properties:
        data:
          type: string
        dataPlaintext:
          type: string
        default:
          type: string
        filename:
          type: string
        repeatableItem:
          type: string
        value:
          type: string
        valuePlaintext:
          type: string
      type: object
    types.AppConfigValues:
      additionalProperties:
        $ref: '#/components/schemas/types.AppConfigValue'
      type: object
    types.AppConfigValuesResponse:
      properties:
        values:
          $ref: '#/components/schemas/types.AppConfigValues'
      type: object
    types.AppInstall:
      properties:
        logs:
          type: string
        status:
          $ref: '#/components/schemas/types.Status'
      type: object
...
    types.Infra:
      properties:
        components:
          items:
            $ref: '#/components/schemas/types.InfraComponent'
          type: array
          uniqueItems: false
        logs:
          type: string
        status:
          $ref: '#/components/schemas/types.Status'
      type: object
    types.InfraComponent:
      properties:
        name:
          type: string
        status:
          $ref: '#/components/schemas/types.Status'
      type: object
    types.InstallAppPreflightsStatusResponse:
      properties:
        allowIgnoreAppPreflights:
          type: boolean
        hasStrictAppPreflightFailures:
          type: boolean
        output:
          $ref: '#/components/schemas/types.PreflightsOutput'
        status:
          $ref: '#/components/schemas/types.Status'
        titles:
          items:
            type: string
          type: array
          uniqueItems: false
      type: object
    types.InstallAppRequest:
      properties:
        ignoreAppPreflights:
          type: boolean
      type: object
    types.InstallHostPreflightsStatusResponse:
      properties:
        allowIgnoreHostPreflights:
          type: boolean
        output:
          $ref: '#/components/schemas/types.PreflightsOutput'
        status:
          $ref: '#/components/schemas/types.Status'
        titles:
          items:
            type: string
          type: array
          uniqueItems: false
      type: object
...
    types.LinuxInfraSetupRequest:
      properties:
        ignoreHostPreflights:
          type: boolean
      type: object
    types.LinuxInstallationConfig:
      properties:
        adminConsolePort:
          type: integer
        dataDirectory:
          type: string
        globalCidr:
          type: string
        httpProxy:
          type: string
        httpsProxy:
          type: string
        localArtifactMirrorPort:
          type: integer
        networkInterface:
          type: string
        noProxy:
          type: string
        podCidr:
          type: string
        serviceCidr:
          type: string
      type: object
    types.LinuxInstallationConfigResponse:
      properties:
        defaults:
          $ref: '#/components/schemas/types.LinuxInstallationConfig'
        resolved:
          $ref: '#/components/schemas/types.LinuxInstallationConfig'
        values:
          $ref: '#/components/schemas/types.LinuxInstallationConfig'
      type: object
    types.PatchAppConfigValuesRequest:
      properties:
        values:
          $ref: '#/components/schemas/types.AppConfigValues'
      type: object
    types.PostInstallRunHostPreflightsRequest:
      properties:
        isUi:
          type: boolean
      type: object
    types.PreflightsOutput:
      properties:
        fail:
          items:
            $ref: '#/components/schemas/types.PreflightsRecord'
          type: array
          uniqueItems: false
        pass:
          items:
            $ref: '#/components/schemas/types.PreflightsRecord'
          type: array
          uniqueItems: false
        warn:
          items:
            $ref: '#/components/schemas/types.PreflightsRecord'
          type: array
          uniqueItems: false
      type: object
    types.PreflightsRecord:
      properties:
        message:
          type: string
        strict:
          type: boolean
        title:
          type: string
      type: object
    types.State:
      enum:
      - Pending
      - Running
      - Succeeded
      - Failed
      example: Succeeded
      type: string
      x-enum-varnames:
      - StatePending
      - StateRunning
      - StateSucceeded
      - StateFailed
    types.Status:
      properties:
        description:
          type: string
        lastUpdated:
          type: string
        state:
          $ref: '#/components/schemas/types.State'
      type: object
    types.TemplateAppConfigRequest:
      properties:
        values:
          $ref: '#/components/schemas/types.AppConfigValues'
      type: object
...
    v1beta1.ConfigChildItem:
      properties:
        default:
          $ref: '#/components/schemas/github_com_replicatedhq_kotskinds_multitype.BoolOrString'
        name:
          type: string
        recommended:
          type: boolean
        title:
          type: string
        value:
          $ref: '#/components/schemas/github_com_replicatedhq_kotskinds_multitype.BoolOrString'
      type: object
    v1beta1.ConfigGroup:
      properties:
        description:
          type: string
        items:
          items:
            $ref: '#/components/schemas/v1beta1.ConfigItem'
          type: array
          uniqueItems: false
        name:
          type: string
        title:
          type: string
        when:
          type: string
      type: object
    v1beta1.ConfigItem:
      properties:
        affix:
          type: string
        countByGroup:
          additionalProperties:
            type: integer
          type: object
        data:
          type: string
        default:
          $ref: '#/components/schemas/github_com_replicatedhq_kotskinds_multitype.BoolOrString'
        error:
          type: string
        filename:
          type: string
        help_text:
          type: string
        hidden:
          type: boolean
        items:
          items:
            $ref: '#/components/schemas/v1beta1.ConfigChildItem'
          type: array
          uniqueItems: false
        minimumCount:
          type: integer
        multi_value:
          items:
            type: string
          type: array
          uniqueItems: false
        multiple:
          type: boolean
        name:
          type: string
        readonly:
          type: boolean
        recommended:
          type: boolean
        repeatable:
          type: boolean
        required:
          type: boolean
        templates:
          items:
            $ref: '#/components/schemas/v1beta1.RepeatTemplate'
          type: array
          uniqueItems: false
        title:
          type: string
        type:
          type: string
        validation:
          $ref: '#/components/schemas/v1beta1.ConfigItemValidation'
        value:
          $ref: '#/components/schemas/github_com_replicatedhq_kotskinds_multitype.BoolOrString'
        valuesByGroup:
          $ref: '#/components/schemas/v1beta1.ValuesByGroup'
        when:
          type: string
        write_once:
          type: boolean
      type: object
    v1beta1.ConfigItemValidation:
      properties:
        regex:
          $ref: '#/components/schemas/v1beta1.RegexValidator'
      type: object
    v1beta1.GroupValues:
      additionalProperties:
        type: string
      type: object
    v1beta1.RegexValidator:
      properties:
        message:
          type: string
        pattern:
          type: string
      type: object
    v1beta1.RepeatTemplate:
      properties:
        apiVersion:
          type: string
        kind:
          type: string
        name:
          type: string
        namespace:
          type: string
        yamlPath:
          type: string
      type: object
    v1beta1.ValuesByGroup:
      additionalProperties:
        $ref: '#/components/schemas/v1beta1.GroupValues'
      type: object
  securitySchemes:
    bearerauth:
      bearerFormat: JWT
      scheme: bearer
      type: http
...
info:
  contact:
    email: support@replicated.com
    name: API Support
    url: https://github.com/replicatedhq/embedded-cluster/issues
  description: This is the API for the Embedded Cluster project.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Embedded Cluster API
  version: "0.1"
openapi: 3.1.0
paths:
...
  /linux/install/app-preflights/run:
    post:
      description: Run install app preflight checks using current app configuration
      operationId: postLinuxInstallRunAppPreflights
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.InstallAppPreflightsStatusResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.APIError'
          description: Bad Request
      security:
      - bearerauth: []
      summary: Run install app preflight checks
      tags:
      - linux-install
  /linux/install/app-preflights/status:
    get:
      description: Get the current status and results of app preflight checks for
        install
      operationId: getLinuxInstallAppPreflightsStatus
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.InstallAppPreflightsStatusResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.APIError'
          description: Bad Request
      security:
      - bearerauth: []
      summary: Get app preflight status for install
      tags:
      - linux-install
  /linux/install/app/config/template:
    post:
      description: Template the app config with provided values and return the templated
        config
      operationId: postLinuxInstallTemplateAppConfig
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/types.TemplateAppConfigRequest'
        description: Template App Config Request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.AppConfig'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.APIError'
          description: Bad Request
      security:
      - bearerauth: []
      summary: Template the app config with provided values
      tags:
      - linux-install
  /linux/install/app/config/values:
    get:
      description: Get the current app config values
      operationId: getLinuxInstallAppConfigValues
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.AppConfigValuesResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.APIError'
          description: Bad Request
      security:
      - bearerauth: []
      summary: Get the app config values
      tags:
      - linux-install
    patch:
      description: Set the app config values with partial updates
      operationId: patchLinuxInstallAppConfigValues
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/types.PatchAppConfigValuesRequest'
        description: Patch App Config Values Request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.AppConfigValuesResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.APIError'
          description: Bad Request
      security:
      - bearerauth: []
      summary: Set the app config values
      tags:
      - linux-install
  /linux/install/app/install:
    post:
      description: Install the app using current configuration
      operationId: postLinuxInstallApp
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/types.InstallAppRequest'
        description: Install App Request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.AppInstall'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.APIError'
          description: Bad Request
      security:
      - bearerauth: []
      summary: Install the app
      tags:
      - linux-install
  /linux/install/app/status:
    get:
      description: Get the current status of app installation
      operationId: getLinuxInstallAppStatus
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.AppInstall'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.APIError'
          description: Bad Request
      security:
      - bearerauth: []
      summary: Get app install status
      tags:
      - linux-install
  /linux/install/host-preflights/run:
    post:
      description: Run install host preflight checks using installation config and
        client-provided data
      operationId: postLinuxInstallRunHostPreflights
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/types.PostInstallRunHostPreflightsRequest'
        description: Post Install Run Host Preflights Request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.InstallHostPreflightsStatusResponse'
          description: OK
      security:
      - bearerauth: []
      summary: Run install host preflight checks
      tags:
      - linux-install
  /linux/install/host-preflights/status:
    get:
      description: Get the current status and results of host preflight checks for
        install
      operationId: getLinuxInstallHostPreflightsStatus
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.InstallHostPreflightsStatusResponse'
          description: OK
      security:
      - bearerauth: []
      summary: Get host preflight status for install
      tags:
      - linux-install
  /linux/install/infra/setup:
    post:
      description: Setup infra components
      operationId: postLinuxInstallSetupInfra
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/types.LinuxInfraSetupRequest'
        description: Infra Setup Request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.Infra'
          description: OK
      security:
      - bearerauth: []
      summary: Setup infra components
      tags:
      - linux-install
  /linux/install/infra/status:
    get:
      description: Get the current status of the infra
      operationId: getLinuxInstallInfraStatus
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.Infra'
          description: OK
      security:
      - bearerauth: []
      summary: Get the status of the infra
      tags:
      - linux-install
  /linux/install/installation/config:
    get:
      description: get the installation config
      operationId: getLinuxInstallInstallationConfig
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.LinuxInstallationConfigResponse'
          description: OK
      security:
      - bearerauth: []
      summary: Get the installation config
      tags:
      - linux-install
  /linux/install/installation/configure:
    post:
      description: configure the installation for install
      operationId: postLinuxInstallConfigureInstallation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/types.LinuxInstallationConfig'
        description: Installation config
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.Status'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.APIError'
          description: Bad Request
      security:
      - bearerauth: []
      summary: Configure the installation for install
      tags:
      - linux-install
  /linux/install/installation/status:
    get:
      description: Get the current status of the installation configuration for install
      operationId: getLinuxInstallInstallationStatus
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.Status'
          description: OK
      security:
      - bearerauth: []
      summary: Get installation configuration status for install
      tags:
      - linux-install
...
servers:
- url: /api